+++ src/index.tsx
@@ -4,4 +4,4 @@ import App from './react/App';
 
 ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
-).render(<React.StrictMode><App /></React.StrictMode>);
+).render(<App />);